requirements.txt
gunicorn
aiodns==2.0.0
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cleo==0.7.6
clikit==0.4.3
cryptography==2.9.2
dill==0.3.1.1
Django==3.0.5
dnspython==1.16.0
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
jeepney==0.4.3
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
msgpack==1.0.0
multiprocess==0.70.9
numpy==1.18.3
openpyxl==3.0.3
pandas==1.0.3
pastel==0.2.0
pathos==0.2.5
pexpect==4.8.0
pkginfo==1.5.0.1
poetry==1.0.5
pox==0.2.7
ppft==1.6.6.1
ptyprocess==0.6.0
pycares==3.1.1
pycparser==2.20
pydnsbl==1.0.2
pylev==1.3.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.8.0
SecretStorage==3.1.2
shellingham==1.3.2
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
tomlkit==0.5.11
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
xlwt==1.3.0
lxml
git+https://github.com/chandusanjith/virustotal2/
psycopg2-binary